% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOrthologs.R
\name{mapIDs}
\alias{mapIDs}
\title{Map moleculer identifiers using ENSEMBL BioMart}
\usage{
mapIDs(DT, ids2convert_col = "target", filters = "hgnc_symbol",
  map_to = "entrezgene", map_to_name = "target_entrezgene",
  biomart_dataset = "hsapiens_gene_ensembl")
}
\arguments{
\item{DT}{data.table containing id to be converted}

\item{ids2convert_col}{column in \code{DT} containing identifiers to be converted}

\item{filters}{Filters (one, in ids2convert_col column) that should be used in the query. A possible list of filters can be retrieved using the function listFilters.}

\item{map_to}{Attributes you want to retrieve. A possible list of attributes can be retrieved using the function listAttributes.}

\item{map_to_name}{how to name the column in the output containing \code{map_to} identifiers}

\item{biomart_dataset}{which BioMart dataset to use for mapping. Full list of possible options: listDatasets("ensembl")}
}
\value{
mapIDs(): data.table \code{DT} containing an additional column (\code{map_to_name})
}
\description{
mapIDs(): Map moleculer identifiers in a data.table using ENSEMBL BioMart - \link[biomaRt]{getBM}
}
\examples{
# load TF regulon data
combined_file = "../regulatory_networks_by_cmap/data/validation_TF_regulons/all_simp_regulons_w_pos_effect.tsv"
regulons = fread(combined_file, header = T, stringsAsFactors = F)
regulons = mapIDs(regulons)
}
